feat: Add SetAgentCard and WithAgentCard methods for custom agent card management
This allows the users of the ADK to define the agent cards as they wish, perhaps in the future I will introduce WithAutomaticAgentCard that is discovered by the defined toolbox, but for now will keep it simple and let the user's of the library define it.

The user's can still generate the skills from the toolbox and create their own GetAgentCard.

Signed-off-by: Eden Reich <eden.reich@gmail.com>