refactor: Remove unnecessary abstractions and add missing methods to an OpenAICompatibleAgent interface
I think it makes sense to be able to get the agent's system prompt from the interface rather than from a private method.

I haven't decided yet whom the toolbox belongs, I think it should be belonging to the agent rather than to the toolbox. A toolbox by itself is not an entity.

Signed-off-by: Eden Reich <eden.reich@gmail.com>