refactor: Remove logging of agent input requests in task handling
This is probably not an ideal implementation and I'm not sure if it follows the Google's A2A specification because the Google's A2A specification doesn't mention any message of kind input_required - it's my invention, not sure if there is another better way to do it. Here I'm giving the LLM to decide if it needs more input.

Signed-off-by: Eden Reich <eden.reich@gmail.com>